Run POYO/POYO+ training in the testing workflow (#62)
Update the CI workflow to include a rule for downloading mc_maze_small, training POYO for 2 epochs, then training POYO+ for 2 epochs (including validation and testing). 

This PR includes adding the configuration files for mc_maze_small.

---

installing `brainsets` is done through cloning the main branch, but in the future (`brainsets 0.1.2` release) we will replace this with a `pip install`